+++ NEWS.md
@@ -2,7 +2,11 @@
 
 ## New functions
 
-* `get_variances()`, to calculate the variance components from mixed models of class `merMod`, `glmmTMB` and `stanreg`.
+* `get_variances()`, to calculate the variance components from mixed models of class `merMod`, `glmmTMB` and `stanreg` (#52).
+
+## Changes to functions
+
+* `find_parameters()` and `get_parameters()` gets a `pars`-argument for `brmsfit` and `stanreg` models, to allow selection of specific parameters that should be returned (#55).
 
 # insight 0.1.2